+++ examples/nostr-shell-example-application/src/integTest/java/org/tbk/nostr/example/shell/NostrShellExampleApplicationShellTest.java
@@ -3,41 +3,23 @@ package org.tbk.nostr.example.shell;
 import lombok.extern.slf4j.Slf4j;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.shell.test.ShellAssertions;
+import org.springframework.shell.test.ShellScreen;
 import org.springframework.shell.test.ShellTestClient;
 import org.springframework.shell.test.autoconfigure.ShellTest;
-import org.springframework.test.annotation.DirtiesContext;
-
-import java.util.concurrent.TimeUnit;
-
-import static org.awaitility.Awaitility.await;
 
 @Slf4j
 @ShellTest
-@DirtiesContext(classMode = DirtiesContext.ClassMode.AFTER_EACH_TEST_METHOD)
+@SpringBootTest(classes = NostrShellExampleApplication.class)
 class NostrShellExampleApplicationShellTest {
 
     @Autowired
     private ShellTestClient client;
 
     @Test
-    void interactiveShellTest() {
-        ShellTestClient.InteractiveShellSession session = client
-                .interactive()
-                .run();
-
-        await().atMost(2, TimeUnit.SECONDS).untilAsserted(() -> {
-            ShellAssertions.assertThat(session.screen())
-                    .containsText("nostr:>");
-        });
-
-        session.write(session.writeSequence()
-                .text("help")
-                .carriageReturn()
-                .build());
-        await().atMost(2, TimeUnit.SECONDS).untilAsserted(() -> {
-            ShellAssertions.assertThat(session.screen())
-                    .containsText("AVAILABLE COMMANDS");
-        });
+    void interactiveShellTest() throws Exception {
+        ShellScreen screen = client.sendCommand("help");
+        ShellAssertions.assertThat(screen).containsText("AVAILABLE COMMANDS");
     }
 }