+++ examples/nostr-shell-example-application/src/integTest/java/org/tbk/nostr/example/shell/command/IdentityCommandTest.java
@@ -3,47 +3,31 @@ package org.tbk.nostr.example.shell.command;
 import lombok.extern.slf4j.Slf4j;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.shell.test.ShellAssertions;
+import org.springframework.shell.test.ShellScreen;
 import org.springframework.shell.test.ShellTestClient;
 import org.springframework.shell.test.autoconfigure.ShellTest;
-import org.springframework.test.annotation.DirtiesContext;
-
-import java.util.concurrent.TimeUnit;
-
-import static org.awaitility.Awaitility.await;
+import org.tbk.nostr.example.shell.NostrShellExampleApplication;
 
 @Slf4j
 @ShellTest
-@DirtiesContext(classMode = DirtiesContext.ClassMode.AFTER_EACH_TEST_METHOD)
+@SpringBootTest(classes = NostrShellExampleApplication.class)
 class IdentityCommandTest {
 
     @Autowired
     private ShellTestClient client;
 
     @Test
-    void testIdentityInteractive() {
-        ShellTestClient.InteractiveShellSession session = client
-                .interactive()
-                .run();
+    void testIdentityInteractive() throws Exception {
+        ShellScreen screen = client.sendCommand("identity");
 
-        await().atMost(2, TimeUnit.SECONDS).untilAsserted(() -> {
-            ShellAssertions.assertThat(session.screen())
-                    .containsText("nostr:>");
-        });
-
-        session.write(session.writeSequence()
-                .text("identity").space()
-                .carriageReturn()
-                .build());
-
-        await().atMost(10, TimeUnit.SECONDS).untilAsserted(() -> {
-            ShellAssertions.assertThat(session.screen())
-                    .containsText("{")
-                    .containsText("\"privateKey\" : \"")
-                    .containsText("\"publicKey\" : \"")
-                    .containsText("\"nsec\" : \"nsec1")
-                    .containsText("\"npub\" : \"npub1")
-                    .containsText("}");
-        });
+        ShellAssertions.assertThat(screen)
+                .containsText("{")
+                .containsText("\"privateKey\" : \"")
+                .containsText("\"publicKey\" : \"")
+                .containsText("\"nsec\" : \"nsec1")
+                .containsText("\"npub\" : \"npub1")
+                .containsText("}");
     }
 }