+++ examples/nostr-shell-example-application/src/main/java/org/tbk/nostr/example/shell/NostrShellExampleApplication.java
@@ -1,14 +1,10 @@
 package org.tbk.nostr.example.shell;
 
-import org.jline.utils.AttributedString;
-import org.jline.utils.AttributedStyle;
 import org.springframework.boot.WebApplicationType;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.boot.builder.SpringApplicationBuilder;
 import org.springframework.boot.context.ApplicationPidFileWriter;
 import org.springframework.context.ApplicationListener;
-import org.springframework.context.annotation.Bean;
-import org.springframework.shell.jline.PromptProvider;
 
 import java.util.Locale;
 import java.util.TimeZone;
@@ -30,10 +26,5 @@ public class NostrShellExampleApplication {
 
     private static ApplicationListener<?> applicationPidFileWriter() {
         return new ApplicationPidFileWriter("application.pid");
-    }
-
-    @Bean
-    PromptProvider promptProvider() {
-        return () -> new AttributedString("nostr:>", AttributedStyle.DEFAULT.foreground(AttributedStyle.MAGENTA));
     }
 }