+++ examples/nostr-shell-example-application/src/main/java/org/tbk/nostr/example/shell/command/IdentityCommand.java
@@ -2,9 +2,9 @@ package org.tbk.nostr.example.shell.command;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.shell.standard.ShellCommandGroup;
-import org.springframework.shell.standard.ShellComponent;
-import org.springframework.shell.standard.ShellMethod;
+import org.springframework.shell.core.command.annotation.Command;
+import org.springframework.shell.core.command.annotation.CommandGroup;
+import org.springframework.stereotype.Component;
 import org.tbk.nostr.example.shell.util.Json;
 import org.tbk.nostr.identity.Identity;
 import org.tbk.nostr.identity.MoreIdentities;
@@ -13,12 +13,12 @@ import org.tbk.nostr.nip19.Nip19;
 import java.io.IOException;
 
 @Slf4j
-@ShellComponent
-@ShellCommandGroup("Commands")
+@Component
+@CommandGroup(name = "Commands")
 @RequiredArgsConstructor
 class IdentityCommand {
 
-    @ShellMethod(key = "identity", value = "Generate a nostr key pair")
+    @Command(name = "identity", description = "Generate a nostr key pair")
     public String run() throws IOException {
         Identity.Account account = MoreIdentities.random().deriveAccount(0L);