+++ nostr-relay-plugins/plugin-allowlist/nostr-relay-plugin-allowlist-db/src/main/java/org/tbk/nostr/relay/plugin/allowlist/db/domain/AllowlistEntryService.java
@@ -38,6 +38,6 @@ public class AllowlistEntryService {
     }
 
     public void remove(XonlyPublicKey publicKey) {
-        entries.delete(AllowlistEntrySpecifications.hasPubkey(publicKey));
+        entries.delete(AllowlistEntrySpecifications.deleteByPubkey(publicKey));
     }
 }